+++ boxsdk/auth/__init__.py
@@ -7,7 +7,14 @@ from .developer_token_auth import DeveloperTokenAuth
 try:
     from .jwt_auth import JWTAuth
 except ImportError:
-    JWTAuth = None  # If extras are not installed, JWTAuth won't be available.
+    JWTAuth = None  # If extras[jwt] are not installed, JWTAuth won't be available.
 from .oauth2 import OAuth2
-from .redis_managed_oauth2 import RedisManagedJWTAuth, RedisManagedOAuth2
+try:
+    from .redis_managed_oauth2 import RedisManagedOAuth2
+except ImportError:
+    RedisManagedOAuth2 = None  # If extras[redis] are not installed, RedisManagedOAuth2 won't be available.
+try:
+    from .redis_managed_jwt_auth import RedisManagedJWTAuth
+except ImportError:
+    RedisManagedJWTAuth = None  # If extras[jwt,redis] are not installed, RedisManagedJWTAuth won't be available.
 from .remote_managed_oauth2 import RemoteOAuth2